still not working so adding "asset importer"

+++ include/Canis/Data/Vertex.hpp
@@ -8,7 +8,7 @@ namespace Canis
         glm::vec3   position;
         glm::vec3   normal;
         glm::vec2   texCoords;
-        glm::ivec4  boneIDs;
+        glm::vec4  boneIDs;
         glm::vec4   weights;
 	};
 }
